refactor: stop using an allocator at `Args.init`
Using an allocator for creating an instance of `Args` is not needed,
I thought I needed, but thinking about it, it's actually not the case.

Why?

Now the function for creating an instance of `Args` is called `Args.create`,
and instead of returning `*Args`, now I'm returning `Args`, this simply
means that the function `Args.create` will return an instance of `Args`,
and this instance will live in the stack-frame of the calling function,
in this case, the `main` function.