refactor: switch to `heap.StackFallbackAllocator`
Currently the allocator being used is a `FixedBufferAllocator` with
a backing buffer of 1KB, this should work for cases where the CLI
arguments total size don't exceed 1KB of memory, but if this limit
is exceeded, an `OutOfMemory` error will occur, which is not good!

Welcome the `StackFallbackAllocator`: this allocator internally uses
a bump allocator of some size, and a fallback allocator, initially, it
allocates memory from the bump allocator, and if the bump allocator
errors with `OutOfMemory`, then it allocates memory from the fallback
allocator.

In this program, the bump allocator has a size of 1024 bytes, and the
fallback allocator is an arena allocator.

The nice thing about this is, if the memory of the bump allocator is enough
for the CLI arguments, then the fallback allocator is never called, and if
the CLI arguments exceed our bump allocator, then the fallback allocator is used.